+++ spec/money/arithmetic_spec.rb
@@ -175,14 +175,13 @@ RSpec.describe Money::Arithmetic do
     end
 
     context 'when conversions disallowed' do
-      around do |example|
-        begin
-          old_default_bank = Money.default_bank
-          Money.disallow_currency_conversion!
-          example.run
-        ensure
-          Money.default_bank = old_default_bank
-        end
+      before do
+        @old_default_bank = Money.default_bank
+        Money.disallow_currency_conversion!
+      end
+
+      after do
+        Money.default_bank = @old_default_bank
       end
 
       context 'when currencies differ' do
@@ -839,14 +838,13 @@ RSpec.describe Money::Arithmetic do
       let(:instance) { Money.usd(1) }
 
       context 'when conversions disallowed' do
-        around do |ex|
-          begin
-            old = Money.default_bank
-            Money.disallow_currency_conversion!
-            ex.run
-          ensure
-            Money.default_bank = old
-          end
+        before do
+          @old_default_bank = Money.default_bank
+          Money.disallow_currency_conversion!
+        end
+
+        after do
+          Money.default_bank = @old_default_bank
         end
 
         context 'and other is money with different currency' do