+++ spec/money/locale_backend_spec.rb
@@ -3,7 +3,7 @@
 RSpec.describe Money::LocaleBackend do
   describe '.find' do
     it 'returns an initialized backend' do
-      expect(described_class.find(:legacy)).to be_a(Money::LocaleBackend::Legacy)
+      expect(described_class.find(:currency)).to be_a(Money::LocaleBackend::Currency)
       expect(described_class.find(:i18n)).to be_a(Money::LocaleBackend::I18n)
     end